requirements.txt
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
click==8.0.1
colorama==0.4.4
decorator==5.1.0
Django==3.2.6
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
iniconfig==1.1.1
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.3.0
pluggy==1.0.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
traitlets==5.1.0
typing-extensions==3.10.0.2
wcwidth==0.2.5
